Add (parts of) InstrumentationFn.
Objects of this class represent the instrumentation functions that are
called from instrumentation hooks. The TESLA version of this code provided
additional helpers for constructing chains of guarded code blocks, but this
code doesn't: we need to think a bit more about how to generalize from the
TESLA use case.