Move policy filename cl argument to OptPass.
The command-line argument specifying the name of the policy file should be
attached to the opt pass we're running rather than the PolicyFile API. When
we're consuming the LOOM API, we don't want policy files to be magically
picked up from the current working directory, etc. This has the side
benefit of making the policy filename visible to OptPass's error handling.