Convert some GEP assertions to conditionals.
Now that we have encountered structure-related GEPs in the wild without two
constant indices (e.g., FreeBSD's `sys/cam/cam_compat.c`, which does a
funny address-of-a-dereferenced-pointer thing), stop asserting that such
GEPs much ahve two constant indices. Instead, take note that these are not
the GEPs we're looking for.