Refactor to support multiple instr strategies.
Previously, we always used the "call out to a function" strategy for our
instrumentation. Even though compilers are clever and can inline things,
we want to be able to explicitly compare the costs of inline vs callout
instrumentation. This commit adds supports for different instrumentation
strategies and re-implements the callout strategy in terms of the new
design.

In addition to enabling a future "inline" strategy, this refactoring also
puts us in a better position to expose the creation of instrumentation to
client code, allowing it to do arbitrary things when instrumentation code
is created (which, depending on the strategy, may not occur every time a
instrumentation *hook* is generated).