Hide Logger's implementation details.
Future Logger implementations (e.g., logging to a file or ktrace) will not
operate as straightforwardly as the existing printf and printf-like loggers.
Consequently, we need to hide some of the details currently exposed by the
Logger class (e.g., `Logger::getFunctionType()`): the new Logger subclasses
won't be able to fulfil these bits of the current contract.

Also create a new SimpleLogger class to hold the implementation bits that
are still common to simple (printf-like) loggers. This class inherits from
the new, streamlined Logger, and {Libxo,Printf}Logger inherit from it.