Add DebugInfo class to parse LLVM debug info.
This class provides useful methods for tracing backwards and forwards
through LLVM Metadata (which is currently only used to store debug info).
This short-circuits the DWARF vs CTF discussion by going directly to the
information that DWARF output is generated from (and, in point of fact,
which CTF could be generated from if we chose to do so).