Make field instrumentation ID-, not name-based.
Now that we have some support for LLVM debug info, we can use it to learn
about the names of structure fields rather than requiring the programmer to
specify the index of the field within the structure. This introduces a
strong dependency on debug information, but it is a much more programmable
model for practical use.