Add generic instruction instrumentation support.
Although most of our instrumentation will benefit from contextual
information about which functions are being called, values are being
changed, etc., it is also desirable to be able to emit generic logs of
the form "an LLVM 'instruction' was executed/evaluated". In particular, this
should allow us to come up with an upper bound on the quantity of
information coming out of a complete process execution.