Uniqify names of generic-instruction instr points.
Instrumentation targets need to have unique names. In the future we should do
something clever with debug information, but in the meantime, just make sure
that every name is unique using the same technique as LLVM's own call graph
printer: take a pointer to the thing that's supposed to be unique and
concatenate that to the name.