Add test for SCCP with degenerate blocks.
When we add instrumentation to otherwise degenerate blocks, the constant
propagation optimizer tends to get cranky: it can no longer fold branches
together (from my current reading of the problem). This test currently
fails, but should provide motivation for the maybe-someday refactoring of
InstrStrategy to stop requiring BasicBlocks to be promoted to "today".