Fix SCCP folding of dead blocks.
When we run constant propagation (as implied by -O2) on instrumented code,
we had been encountering problems with dead blocks that couldn't be
eliminated. It turns out that we had been trying to pass BasicBlock operands
directly to instrumentation functions such as xo_emit... this isn't flagged
as evil by LLVM's IR verification pass, but it is a Distinctly Bad Idea.

This commit applies a simple fix (and test updates): just ignore BasicBlock
operands when invoking instrumentation!