Split 'instrument-everything' into two tests.
This test can be instrumented on any host platform, but the result can only
execute on platforms that match the target triple (i.e., freebsd). In order to
test as much as possible in as many places as possible, split the test into two:
one that can run anywhere (checking the instrumentation) and one that can only
be run on FreeBSD (checking the output of the built program).