Add instrumentation block structure option.
We have traditionally used an instrumentation structure oriented around
BasicBlocks (injecting them inline or in instrumentation functions), but
for inline instrumentation in particular, we might prefer to keep things
very simple and just inject instructions.

This commit gives us the option: it doesn't add the Policy flag to
control how the option is used.