Add class diagram for Loom.
Add a description of the Loom class hierarchy and a script to graph it
(using the umlgraph tool). The API description is using Java syntax as a
rough model of the actual Loom API. This is useful for presenting the
big picture simply, but deeper inspection will require Doxygen together
with code reading.