Stop trying to emit C-style strings.
Sometimes an i8* is a string, but sometimes it's not. Such pointers,
when derived from GEPs of static strings, are often safe, but we can't
be sure. Emitting i8* as a string works for bsdgrep, but not for cp,
which causes a void* from mmap be casted to an i8*. Furthermore, we
expect to encounter even more danger once adversaries get involved.

Perhaps we'll find a way to safely deal with C-style strings in the
future, but for now, take the conservative approach of treating them
like all other pointers: emit the address, not the content.