Improve error message when llc not found.
When the current PATH does not contain `llc`, provide more explicit
suggestions to the user than the normal Python backtrace. That backtrace
contains a buried "No such file or directory" but doesn't state clearly
which file was not found, leading to the impression that `lit.cfg` is
the missing file when it is, in fact, not missing.