Wrote a new logger that for stderr_print
Wrote a new logger that works exactly the same as printf except that it
call stderr_printf. Also added an option for this in the policy file.

This uses a stderr_printf function that will need to be linked in. The
function should simply call vfprint with the arguments and stderr.