Commented out adding  attributes to callout functions
When adding callout function, Loom would check whether the previous
instruction was a call and add that call's attributes to the callout.
This seems ill thought-through: there are many attributes that cannot
just be transfered in this way: the example that led me to comment this
out was one which referred to an argument (the callout function had
none). Maybe inform the CADETS team of this.