Adds support for initialization of loggers where needed.
Some loggers require initiallization to occur. These changes allow a
Logger to declare it need initialization and define the initialization
code. This code will be added to the beginning of main() in full
program instrumentation.

When instrumenting partical programs or libraries where main() is not
expossed to the intrumenter, this technique will not work.