Uses argument types to construct instr call for variadic functions.
When instrumenting the call of a variadic function we cannot use the
function parameters to build the instrumentation call as it will not
account for all arguments passed. Instead we build the function using
the types of the arguments.

This can cause an issue with fucntion overloading that llvm doesn't
handle well. If there are two functions with the same number of
arguments but different type signatures, getOrInsertFunction fails. To
avoid this, all callout functions for variadic functions include the
pointer address from the call in the name.